pdfwrite - clean up proerly after errors writing TrueType fonts
OSS-fuzz #51745

Commit bd08091d295effb51de2cc7e739870e7b2628e38 already tackled a number
of other cases of the same problem, but sadly missed this one. Once
we've begun writing a font file we must end it before returning or we
will end up with the wrong stream in the pdfwrite device structure
which will lead to all kinds of problems.

Add code to handle the missing case here.