Overprint, transparency, alpha buffer, stroke fill interaction
This spins out of bug 705260.

The main focus of this commit was to get the alphabits stroke
fill operation to work properly with the pdf14 device.  When
fill stroke operations occur with overprint and transparency
there are specific pdf14 setups required as outlined by the pdf
spec.

Prior to this, the pdf14 stroke/fill method would do all the
required setup within that method.  This setup needed to be
accessible to the abuf device when it is doing its stroke/fill
method. This was achieved through the use of a special operation
which has been renamed from gxdso_overprint_op to gxdso_abuf_optrans
to make it clear that it is really only used by the abuf device
for communicating to the overprint and transparency compositors.
Essentially the parts that handled all the special transparency
operations were pulled out of the pdf14 stroke/fill method and made
available for use during the special operations call of
gxdso_abuf_optrans.  A state variable was maintained to pass information
as we move from fill to stroke.

There were several additional issues found during this work.
1) The pdf14 op_state which tracks if we are doing fill or
stroke was not getting updated in pdf14_set_marking_params
which caused the wrong overprint state to be used when we
had stroke/fill operations with differing overprint settings
during clist reading.
2) The alpha buffer device must clear its buffer between
the fill and stroke operations and before the overprint state
is restored to its prior state. Otherwise data can remain in
the memory buffer and is not released until the device is closed,
at which point the overprint state may have changed.
3) From careful review, it was clear that when the stroke
and fill alpha both equal 1 and the blend mode is normal
no special groups are pushed in the presence of overprint.
4) The interpreter was pushing a group during stroke/fill
overprint cases. This has been removed as the pdf14 device
will handle all of these decisions.

In case git messes up the attribution, this code was originally
written by Michael Vrhel, and then the gxdso behaviour was
altered by Robin Watts. Credit to Michael, blame to Robin as
usual!