Handle CumulativePageCount not being defined
In order to preserve a consistent page count when using multiple input PDFs
the PS code defines and maintains CumulativePageCount value.

This is initially setup in newpdf_runpdfbegin.

But to gain finer grained control over the interpreter, it's necessary to skip
using newpdf_runpdfbegin, meaning newpdf_gather_parameters will throw an
undefined error on CumulativePageCount.

So tweak newpdf_gather_parameters so spot when it isn't defined, and set it to
one.