Graphics library - better fix for HexDecode filter
OSS-fuzz #51789, #51981, #51991, #51992

Commit 9e1a0aa79da04c1ccf6bf6e030c551d03149bb3e fixed OSS-fuzz 51789
but it unfortunately left more problems open.

To be honest the real problem is the incomprehensible nature of the
stream code, but in this case there was some documentation which I
missed. s_hex_process() documents what the return codes for the string
interface mean. The commit changed the return code from 0 (ran out of
input) to 1 (filled the string) which fixed the problem but really
for the wrong reasons.

In this commit; revert the change in behaviour of s_hex_decode() so
that it follows the comments. Instead, in s_AXD_process() if we have
a dangling half-byte, when we get a return value of 0 from s_hex_decode,
reset ss->odd after storing the half byte in the buffer so that when we
next return here we will drop straight into case 1, when the return
value is 0 from s_hex_decode.

This solves the original problem and all the problems stemming from the
first commit (so far anyway). Cluster is still happy with this.