pdfwrite - fix Permissions on 64-bit Linux build
Bug #705951 "-dPermissions=<negative>: 32bit gs -> /P <negative>, 64bit gs -> /P <unsigned int>"

The code in pdf_close to write out the Encrypt dictionary was writing
the Permissions as a long, which is fine on Windows where it is always
32-bit, but doesn't work correctly on Linux because long is 64-bit on
a 64-bit build on Linux.

Change the print to use an int instead of a long, and also alter the
structure member to be a 32-bit integer (not required but might as well
be consistent).