Bug 705939: Fix BaseEncoding logic for fonts flagged symbolic
For a font encoding defined in a PDF dictionary (i.e. with a Differences array)
for a symbolic font, we were ignoring the BaseEncoding entry, and opting to use
the font's native encoding.

In fact, we should always use the BaseEncoding if it exists, and fall back to
the native encoding only if BaseEncoding isn't specified.

With non-symbolic fonts, we fall back to StandardEncoding in the absence of a
BaseEncoding entry