Bug 705931: Avoid double encoding filenames with gs under gpdl.
When the postscript interpreter is run by itself, it has to
cope with decoding arguments supplied on the command line
into our internal format of utf8. For windows, this means
that it has to cope with converting from 'local' encoding to
utf8.

When we run gs under gpdl this decoding is done under the
pl level, so it's important that gs does not do such decoding
otherwise it will 'double decode'. Accordingly, disable the
decoding in the psitop.c layer.