GhostPDF - wrap streams in SubFileDecode filter
The Flate decode filter was throwing errors with pdfi that did not
occur with the old interpreter. On investigation this turns out to be
because the old interpreter wraps streams in a SubFileDecode filter.

The Flate filter (zlib interface) does not throw an error if it is
unable to read the final 4 bytes of the input in order to generate and
and verify a checksum. The streams in this case were all short by one
byte, so the filter ignored the error.

However pdfi wasn't wrapping the stream in a SubFileDecode and read on
past the end of the stream. Thus returning enough bytes to satisfy the
checksum read, but obviously not reading a valid checksum. This *does*
cause an error, but one we deliberately ignore, apart from printing a
warning.

Having done some checking, it seems that wrapping streams in a SubFile
filter doesn't add too much to the processing time, and it prevents us
running off the end of streams. So this commit does exactly that.

If the stream length we have is valid we use that as the EODCount, if
the stream length is not valid, then we supply 'endstream' as the
EODString instead. This will be slightly slower but the file is already
technically invalid anyway.