Graphics library - change error output to a debug output
We specifically ignore checksum errors on FlateDecode streams because
(allegedly) Adobe, Apple and xpdf do. Note we do not differentiate
between PostScript and PDF input!

Although we don't throw an error, we do emit a message, and that message
was written to stderr. This seems wrong, because we are not treating
this as an error. In addition we cannot honour either -q or -dQUIET
in the stream code, because that is all handled back up in the
interpreter, so there's no way to suppress this (not very helpful)
message.

Changed to use dmprintf1 so that the message only occurs on a debug
build.