GhostPDF - improved warning messages with trailer dictionary problems
For historical reasons any problem with the trailre dictionary or any
of its contents was reported as a 'bad trailer dictionary'. This could
be slightly misleading; for example if we process only part of a PDF
file which contains Outlines, to the pdfwrite device, then it any
outline with a Destination outside the final page range would trigger
a warning, which would appear as a bad trailer dictionary.

Just add a bunch more warnings to cover all the possible problem
areas.