Printer memory device not replace device special op
In gdev_mem_set_planar_interleaved only set the special op
to the memory device if it is not already set.  This is
done to ensure the "real" printer target special op will
be called when we are running in immediate mode.  Can
be replicate for example by running

-sDEVICE=psdcmyk -o output.psd -r72 -f ./examples/tiger.eps

and seeing that the psdcmyk special op is never called.