GhostPDL + GS - do not make PS colour spaces into PDF ones
This was reported (out of Bugzilla) by Robin Watts, noticed in passing
while debugging a different problem. Using the bug file (bears.pdf)
from Bug #705975 and the following command line:

-r1500 -sDEVICE=tiffsep -o /temp/out.tif /temp/bears.pdf

if the output file out.tif cannot be written (read-only or held open by
another application) then this caused an error. On a Memento build on
Windows it causes a corrupted block on exit, on Linux it was giving a
warning about attempting to free an unknown PDF object type.

In all cases the error occurred during the PostScript end of job restore
when there should be no PDF objects and the PDF interpreter context had
been released.

The problem was due to colour spaces, when we create a colour space in
PDF we set a callback, called when a colour space is freed, which does
extra work specific to the PDF interpreter. However we were ending up
adding these callbacks to colour spaces allocated by the PostScript
interpreter.

This comes about because gs_setcolorspace() does not alter the current
colour space if we attempt to switch to the same space. So in this case
the current colour space in the PostScript interpreter was DeviceGray
and the PDF interpreter then set DeviceGray, which did not result in a
new colour space in the graphics state. But we went ahead and assigned
the colour callbacks assuming that it had. When we finally freed that
space we would then attempt to handle it as a PDF space, but the PDF
context had long gone, resulting in an error.

To fix this we now check the colour space after calling gs_setcolorspace
to see if it has changed. If it has not we do not assign the PDF colour
callbacks; either we have previously done so (if its a PDF space) or it
is a PostScript space and we absolutely do not want to set the callbacks.

Why does this not show up normally ? Because we do a showpage after
every page, which does an initgraphics, which resets the PostScript
colour space, and at this point the PDF context is still valid. In order
to get this problem we currently need 'showpage' to return an error.

But it is still wrong, much better to avoid the situation.