GhostPDF - preserve current colour around transparency groups
Bug #705966 " Problem with transparency of a white box turning black (PDF exported from LibreOffice)"

The PDF file has a Form XObject which fills a rectangle but does not
set the colour or colour space before doing so; it 'inherits' the
colour and colour space in force at the time the Form is executed.

However, the Form also has a transparency Group, and that Group has a
/CS (colour space) which differs from the parent.

We were preserving the current colour space by retrieving the space
before calling pdfi_trans_begin_form_group(), which sets the colour
space before calling gs_beging_transparency_group(), and the setting
the space back after pdfi_trans_begin_form_group, but we were not
preserving the color value(s).

This commit additionally retrieves the current colour, and modifies
pdfi_form_execgroup() to take a client_color as well as a pointer to a
colour space, and set both the space and the colour before running the
Form content stream.

This retains the colour and space at the time the form was executed
while still setting the colour space correctly for the Group.