xpswrite - free XPS image enumerator on error
OSS-fuzz 52492

The xps_begin_typed_image function allocates an xps_image_enum_t
structure, which it returns to the called in 'pinfo'. However, if
an error occurs it neither updates pinfo nor frees the enumerator
itself.

When the input is PostScript this is adequate (if poor) because the
enumerator is garbage collected. For non-GC interpreters this leads to
a memory leak. for GhostPDF inside Ghostscript it causes a crash
because the enumerator includes a pointer to a colour space, which is a
PDF colour space. Because the enumerator is effectively orphaned it is
not freed until the end of job restore, whereupon it releases the colour
space, which tries to do the extra work required by the PDF interpreter.

But by this time the PDF interpreter itself has gone, so we get a crash.

Fix it by having the function properly free its memory on error instead
of just letting it leak. While we are here, fix the various places the
simply return gs_error_xxxx to use the return_error macro.