GhostPDF - detect integer overflow when parsing numbers
This stems from OSS-fuzz 52535

Part of the problem is that an image in the PDF file has a /Height which
is -2103029590437764, which breaks even a 64-bit value. We did not
detect that, but carried on, which resulted in the number becoming
negative, before we had applied the negation from the '-' sign.

Overall this results in a very large negative number becoming a very
large positive number.

The xpswrite device uses libtiff to write images into XPS files as TIFF
format images. The libtiff library tries to allocate enough memory to
hold the entire image in memory which, with the broken Height value,
results in trying to allocate many terabytes of memory, which causes
the error. We've tried getting the libtiff maintainers to provide us
with a hook so that we can use our memeory manager instead of the
system malloc, on two occasions providing patches to do so, but have
been rebuffed. Obviously if we cannot use our memory manager then we
cannot apply our limits to memory use.

This commit does not address that, it detects the integer overflow
and clamps the value to the last digit before it overflows. Arguably the
value could be reset to 0 but we have no guidance in the spec on this so
clamping it is as good as zeroing it.

For this specific case it solves the problem, because the Height remains
negative and we don't try to render image with Height less than 1.

While it doesn't solve the general case its worth having the overflow
detection anyway.