GhostPDF - detect failed ICC colour spaces
From OSS-fuzz 52552 - command line:

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dDOINTERPOLATE -dNOMEDIAATTRS  /bugs/oss-fuzz/52552.pdf

This exhibits as a floating-point exception in the image scaling filter
caused by -DDOINTERPOLATE, but the actual problem is that the colour
space has zero components, leading to a divide by zero error.

The actual problem is when setting an ICC colour space, the PDF
interpreter was missing an error condition (expected number of colour
components being 0).