oss-fuzz 52636: Check ref type before using
For Type42/TTF fonts we store the sfnts data in an array of strings, which the
FAPI code "seeks" through to read the font data. In this case, one of the
elements in the array is an integer rather than string, and since we weren't
checking the ref type, using is as a string would crash.

Add a check for the ref type, and throw an error if it's not a string.