GhostPDF - cater for a Dests Name tree
Bug #705966 " New PDF interpreter disables link jumping from table of contents"

Link annotations require a /Dest which can be specified either as a
direct entry, or as a named destination. Named destinations can be
found in the Catalog as either a Dests dictionary (PDF 1.1) or a
Dests entry in the /Names dictionary (PDF 1.2+). If we have a Dests
entry in the Names dictionary then it can be either a flat tree (there
is a /Names entry in the root node) or a real tree (the root node has
no /Names but does have a /Kids).

The code for dealing with Dests was not catering for real trees. This is
clearly an oversight because we have the code to deal with them already
for Goto destinations in /Action annotations. Leverage the same code
for Link annotations.

At the same time.... If we don't have a named destination, we don't
have a /Dests in the Catalog, and we don't have a /Names entry in the
Catalog, return an 'undefined' error. Fix a couple of incorrect strings
used for gs_alloc_bytes.

Add some circular reference detection/protection in the existing tree
processing code. This is an omission which it is important to rectify,
sooner or later the OSS-fuzz tests would find this.