GhostPDF + GS - Fix negative llx and lly Box with FitPage
Bug #705992 "Converting PDF to TIFF with ghostscript causes offset output when using -dFitPage and -sPAPERSIZE options"

The files supplied with the report all have a MediaBox (and other Boxes)
where the origin is not only not 0,0 but is in fact negative. A
possibility I had overlooked previously.

This commit fixes that.

This does lead to some changes in behaviour when compared to the old
PDF interpreter.

The output from the old interpreter was inconsistent with 'landscape'
pages, as evidenced in the bug report. In fact these pages are not
landscape at all, in the sense that the media and content is all defined
in portrait. The reason they are displayed landscape is because the page
contains a /Rotate key with a value of 270.

When the output was to a rendering device, such as TIFF, the landscape
files were sized to fit on Legal media (to satisfy -sPAPERSIZE=legal and
-dFitPage) BUT the output was then rotated, resulting in a landscape
bitmap.

When the output was a high level device, such as pdfwrite, then the
landscape files were written as portrait files, sized for Legal media
and rotated so that the content is 'upside down'.

This also happened with the rendering devices when using
-dDEVICEWIDTHPOINTS and -dDEVICEHEIGHTPOINTS to set the media size to
Legal.

When using -dDEVICEHEIGHTPOINTS and -dDEVICEWIDTHPOINTS to set true
landscape media, and fitting the page to that, the rendering devices
coped correctly, but the pdfwrite output has a badly sized CropBox
leading to the files displaying totally incorrectly in Acrobat.

The new code is consistent across all devices, rotations, orientations
of both actual media and requested media.

The requested media in the bug report is Legal, and Legal media is
portrait, not landscape, so it seems to me that producing a landscape
bitmap is wrong. We now produce portrait legal output. The pdfwrite
device also produces portrait media output (as before) but the content
is the correct way up, not rotated.

To achieve landscape output, you need to specify landscape media (which
can't be done with -sPAPERSIZE as only 11x17 has a defined landscape
version), by using -sDEVICEWIDTHPOINTS and -dDEVICEWIDTHPOINTS and
setting -dFIXEDMEDIA to prevent the media size being changed to match
the requested input.