Tweak ps2writes image capturing devices.
These devices are used for capturing images in formats that are
not normally supported by PS (such as rendered shadings).

The device creates a memory buffer of a given width/height (w, h),
with a 'mapped' offset (x, y) from the origin. Writes into the
device are then adjusted by the mapped offset so that the device
ends up capturing the bbox from (x, y) to (x+w, y+h).

Now, we'd very much like it so that callers can spot that it's
not necessary to call devices with 'out of range' writes. i.e.
a caller should know that it's safe to avoid calling a write
into a device that doesn't mark the (0,0) to (dev->width, dev->height)
region (for devices with non-zero width/heights at least).

These devices break that.

So... we change these devices here so that when viewed from
outside, their width/height includes the given offset. The
fill routines for these devices correct the width/height
while actually doing the fill, so the internals still see the
correct area.