Color replacement update
Allow color replacement calls to make it through the pdf14
device to the target device even when the color spaces
do not match. It is up to the target device handler of
gxdso_replacecolor to deal with the setting of the "device"
color, which in this case is going to be put into the
pdf14 device which may have a different color space
than the target device. To make it
easier for the target device handler to know that this
is occurring, pdf14_iccprofile is passed in the special
operation allowing the handler to know what the pdf14's
current device color space is.  Right now, psd_spec_op
as an example will only deal with this if the pdf14
color space is not gray or rgb.  It is possible to
deal with it in those cases, but that is up to whoever
is writing the replacement code and is beyond what is
needed for the demo. Also, I added code
to invert RGB fills when ENABLE_COLOR_REPLACE is set in
the psdcmyk devices as an example.