GhostPDF - fix an oversight processing CMaps
File supplied by customer, not made public. Email subject line RSP-11363

The PDF file uses 4 different CIDFonts and does not include any of them.
All the fonts have Identity Encodings, CMap and CIDToGIDMap entries
which essentially means they will only render correctly if exactly the
right font is used as a substitute.

However, the fonts also contain ToUnicode CMaps, and when substituting
fonts we can use that extra information to try and retrieve an
appropriate glyph from the substitute font. For the majority of the text
in the document this works fine, but for a small number of glyphs, all
with caron accents, they do not render at all. This is because they are
specified individually, not as ranges, and have the top byte of a 2
byte Unicode value non-zero.

Eg:
<0122><0122><011B>

Due to an oversight in the CMap parsing of individual entries, only the
lowest value byte was being considered. Fix that here.

This makes 3 Sumatra test files render differently; in two cases we
simply render a different kind of garbage (the files are broken, badly)
in the third case we get a progression.