GhostPDF + GS - fix ShowAnnotTypes and PreserveAnnotTypes
Bug #706036 "-dPreserveAnnots=false disregards /ShowAnnotTypes []"

There were a couple of problems here. Firstly the switches were
mis-spelled as SHOWANNOTTYPES and PRESERVEANNOTTYPES in both pdf_main.ps
(passing the values to the PDF interpreter) and in the PDF interpreter
itself. This meant that neither switch was being processed.

In addition the code to read the values was incorrectly assuming that
the value was a 'list' of strings. This is not the case for either
control, and indeed simply isn't possible in PostScript. In both cases
the value should be an array of names.

The code now accepts an array of either names or strings for both keys
from PostScript.