GhostPDF - Treat an invalid /Type for Catalog as if it were missing
Bug #706038 " Can't read some PDF files with the new PDF interpreter: "Couldn't initialise file" / object lacks an endobj"

The PDF file her is invalid; the Catalog dictionary has a /Type of
'/Calalog' instead of '/Catalog'. The Type is a required entry and must
be /Catalog to be valid.

However we already treat a missing /Type as an error but continue if the
purported Catalog dictionary contains a /Pages key (the only other
required key). So do the same here; if the trailer dictionary points us
to a dictionary which has a /Type which is not /Catalog, but it does
have a /Pages key whose value is a dictionary, then treat it as a valid
Catalog.

We do, of course, issue an error message at the end of processing the
file.