OSS-fuzz 52879 "Global-buffer-overflow in get_unpack_proc"
The PDF spec allows (more craziness) a JPX image in a PDF file to have
a different colour space to the declared space in the PDF file, and the
JPX should take precedence.

In this case the JPX image has been fuzzed leading to it declaring that
the image has 48 bits per component, which is outside the legal values
for PDF. Although the PDF interpreter does check the BPC, it does it for
the image data in the PDF file, which has been overridden by the data
in the JPX image.

So here we check the BPC value after potential overrides. We also add
a check in get_unpack_proc to make sure we don't over or under run the
array of procedures and finally we add a check in the XPS image writing
code to check the returned decode procedure and if it is NULL to exit
with an error (the only other place that uses this, the clist, already
exits cleanly if the returned proc is NULL)