Memento: Make showDetailedBlocks gate the block listing.
The MuPDF version of Memento was using a newly added
memento.verbose for that, but we're already using memento.verbose
for something different here.

Merging the two "forks" of Memento together, something has to
give, but this way at least Julian (I believe it was him
that wanted this) can still suppress the block listing
by using showDetailedBlocks instead of verbose.