pdfwrite - fix bounds check in type 2 (CFF) font interpreter
OSS-fuzz 53054 "Stack-buffer-underflow in gs_type2_interpret"

The bounds check at case ce3_hflex1 only checked back to csp - 3, but
the code here accesses back as far as csp - 7.

In addition; at the flex: label (which this code gets to via a goto) we
access as far back as csp - 12 (and ce2_flex chacks csp - 12) but the
bounds check in ce2_hflex only goes back to -11 (we add 4 to csp
so -7 turns into -11).

Similarly at ce2_hflex checks back to csp - 5, and adds 6 to csp which
again looks like it is potentially accessing off the bottom of the
buffer.

So fix those cases as well before OSS-fuzz can find them.