GhostPDF - Preserve UserUnit when the output device supports it
Bug #706059 "PDF interpreter does not preserve UserUnit with pdfwrite"

The bug report is in two parts; the first is that pdfwrite does not
automagically insert UserUnit and alter the scaling of files when the
media size exceeds the Acrobat architectural limitation of 14,400. We
do not regard this as a Ghostscript bug, since the PDF format does not
limit the MediaBox, this is a limitation of Acrobat.

The second part correctly noted that the new PDF interpreter was not
preserving any UserUnit from the input file when the output device was
pdfwrite (the only device which supports preserving UserUnit).

This commit preserves any extant UserUnit (and does not scale either the
media or the content) when the output device supports UserUnit and the
-dNoUserUnit switch isn't set. This now works for the case when the PDF
interpreter is built into Ghostscript and when it is a stand-alone
binary.

A couple of files in the test suite show very slight (pixel level)
changes now, because we preserve the UserUnit and so the scaling is not
quite the same.