GhostPDF - simplify InkList annotation appearance generation
OSS-fuzz 53179

The fuzzed file has a bad initial co-ordinate:

/InkList [[124.666666666 66676769.3333333333334 124.66666666666667 768.6666666666666

The way we draw InkList annotations, this causes us to try and draw a
Bezier curve and because the second co-ordinate is unaffected this
becomes a truly huge curve segment, too long to cope with. So we
subdivide it and try again, still too big, subdivide and try again....

We end up subdividing many times, producing many smaller curves which,
later, will turn out to be off the page and not rendered. But before we
can get to that point we spend an awful lot of time chopping up curves.

As long as we use curve segments there really isn't anything we can do
about that, but.... There is no reason we have to use curve segments
rather than straight lines. The spec says

"When drawn, the points are connected by straight lines or curves in
an implementation-dependent way"

and points us to implementation note 93 in Appendix H (1.7 spec) which
says:

"Acrobat viewers always use straight lines to connect the points along
each path."

In fact, looking at the display in Acrobat X, I don't believe that
statement is true, but it allows us the option of doing this. If we do
that instead of using curves we don't have to subdivide them and so we
can run the file *much* more quickly.