oss-fuzz 52986: pdfwrite - free copied font on error
During the process of copying a font for pdfwrite/ps2write, if we encounter an
error before the copied font is completed, we'd skip to the "fail" code
before assocating the copied font with the pdfwrite font object, meaning it
relied on garbage collection to clean it up (never safe in devices!).

Explicitly free the copied font, before skipping to the "fail" code - later
error conditions, the copied font is already linked to the pdfwrite font,
and thus cleaned up correctly with that.