pdfwrite - fix page offset hints table for Linearized PDF
This started off with bug #706078, although the reporter did not
specifically mention this problem, and pdfwrite has always had this bug,
we'll fix it here.

The linearisation code was not updating the usage record for the 'final'
page in the output file. If that happened to be the only page then we
could write an invalid number to the page offset hints table, claiming
that the least number of objects on the page was 0, which it cannot be
because we include the page object, so there must be at least 1 object.