GhostPDF - check for self-referencing named ColorSpaces
OSS-fuzz 53393

The fuzzed file corrupts a named ColorSpace in the Resources dictionary
of the page. The name should be /DefaultRGB with an indirect reference
to the space, what we actually get is :

/ColorSpace<</De////////faultRGB 3 0 R>>

The odd characters are 0x0F bytes. So that defines a named ColorSpace
/De as being the name <0x2f0x0F> and then defines the named space
/<0x2F0x0F> as being /<0x2F0x0F>, ie it is self-referencing.

We detect that for indirect objects, when dereferencing them, but
there's no convenient central place to do that for self-referencing
names, we just have to watch out for it when trying to get the object.

Added a new error to cater for this case.