graphics library - don't try to fill zero-area clips
OSS-fuzz 52884

The file tries to render a shading into a clip which has zero area. The
shading is itself corrupted, which leads to us not being able to handle
the points using fixed-point arithmetic, so we end up decomposing it
into many, many subsidiary patches, none of which actually mark the
page (obviously because they are clipped).

This just short-circuits that leading to much faster completion.