GhostPDF - handle dictionary Dests in Names tree.
Confidential files supplied by customer, no bug report.

This is a follow-on from commit 073c1adebeef38317ae0e6cb0f6d366ddaadb38a
(fix for bug 705966) which added the missing ability to deal with Names
trees for named destinations.

That code expected the Dest to be an array, but it is possible for it to
be a dictionary with a /D key (whose value is an array). Deal with that
here.