PostScript interpreter - Use the correct ref for DeviceN /All ink
File from customer, no bug report.

We have special processing for the /All ink in a DeviceN colour space,
and that processing was assuming that the ink name was a name object
which it then wanted to turn into a string object in order to get the
actual string representing the name.

The file in question, however, defines the names array (quite legally)
as:

[(All)]

So the ink name is a string, not a name. The code explicitly used
name_string_ref to turn the name into a string, which unsurprisingly
failed because the object isn't a name.

In fact we have already dealt with this earlier in the code, all we need
to do is use the correct variable and it all works correctly.