GhostPDF - minor change with broken named destinations
Follow-on from commit1042469a5225ff063c465e61dbd3ebb50c770006, the
customer didn't like the fact that we were dropping the Link annotations
from the document.

This happened because the value associated with the /Dest key was neither
a dictionary nor an array (as it is specified to be), but was a null
object. Which caused an error, so we dropped the annotation.

Previously (old interpreter and prior versions of pdfi) we retained the
Link annotation and dropped the Dest, which still makes it invalid of
course....

This commit simply restores that behaviour, while noting the error for
reporting at the end.