Use a 5 point tolerance with NupControl to decide when to flush the nest.
Bug 705970 had a first page which was slightly different dimension compared
to the next two pages, resulting the the first page being on a master page
by itself, and the next two pages nested side by side (without FIXEDMEDIA).

The tolerance of 5 points was chosen because that is the tolerance for PostScript
media size matching.