GhostPDF, N-up and pdfwrite - don't preserve some pdfmarks with N-up
Bug #705969 "Nup fails with pdfwrite"

The N-up device works with the combination of pdfwrite and the old PDF
interpreter by disabling all pdfmarks. This is sub-optimal, because a
number of pdfmarks don't rely on the page order or media size (eg Stamp,
Text and other marking annotations).

This also didn't work with the new PDF interpreter at all, because it
doesn't use the PdfmakCapable special op. Which meant that all pdfmark
operations were continuing to the pdfwrite device, in particular the
various Box values (CropBox, ArtBox, etc), and the inclusion of those,
which depend on the MediaBox being preserved, caused the output to be
incorrect

So we create two new device special ops; ModifiesPageSize and
ModifiesPageOrder. If a device returns true for ModifiesPageSize then we
no longer pass the Box values (because they won't be correct. If a
device returns true for ModifiesPageOrder then we don't send Outlines
or Dests, because these can depend on the Page ordering.